+++ lib/access_token_agent/unsupported_token_type_error.rb
@@ -1,5 +1,5 @@
 module AccessTokenAgent
-  class InvalidTokenTypeError < Error
+  class UnsupportedTokenTypeError < Error
     def initialize(token_type)
       super("Expected token_type to be 'bearer', but was '#{token_type}'.")
     end