+++ spec/access_token_agent/token_spec.rb
@@ -3,7 +3,13 @@ require 'spec_helper'
 module AccessTokenAgent
   describe Token do
     subject { Token.new(auth_response) }
-    let(:auth_response) { { 'token_type' => 'bearer', 'expires_in' => 3600 } }
+    let(:auth_response) do
+      {
+        'token_type' => 'bearer',
+        'expires_in' => 3600,
+        'access_token' => 'test'
+      }
+    end
 
     describe '.valid?' do
       context 'when token has not expired' do
@@ -11,7 +17,13 @@ module AccessTokenAgent
       end
 
       context 'when Token has expired' do
-        let(:auth_response) { { 'token_type' => 'bearer', 'expires_in' => 0 } }
+        let(:auth_response) do
+          {
+            'token_type' => 'bearer',
+            'expires_in' => 0,
+            'access_token' => 'test'
+          }
+        end
         it { is_expected.not_to be_valid }
       end
     end
@@ -23,10 +35,29 @@ module AccessTokenAgent
       end
 
       context 'when token_type is not supported' do
-        let(:auth_response) { { 'token_type' => 'mac', 'expires_in' => 3600 } }
+        let(:auth_response) do
+          {
+            'token_type' => 'mac',
+            'expires_in' => 3600,
+            'access_token' => 'test'
+          }
+        end
+
+        it 'raises an UnsupportedTokenTypeError' do
+          expect { subject }.to raise_error UnsupportedTokenTypeError
+        end
+      end
+
+      context 'when access_token is missing' do
+        let(:auth_response) do
+          {
+            'token_type' => 'bearer',
+            'expires_in' => 3600
+          }
+        end
 
         it 'raises an InvalidTokenTypeError' do
-          expect { subject }.to raise_error InvalidTokenTypeError
+          expect { subject }.to raise_error MissingAccessToken
         end
       end
     end