+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 - Deprecate the `authenticate` method in favor of the new `token` method
 - Allow to configure from which path to get the access token
 - Put all errors into the AccessTokenAgent namespace
+- Actually return a token when faking auth
+- Rename error raised for unsupported token types
+- Ensure that access token response carries an access token
 
 ## 3.1.1